requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
boto3==1.26.30
botocore==1.29.30
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
cssselect==1.2.0
dnspython==2.2.1
fake-useragent==1.1.1
idna==3.4
importlib-metadata==5.1.0
jmespath==1.0.1
lxml==4.9.1
parse==1.19.0
pyee==8.2.2
pymongo==4.3.3
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
requests-html==0.10.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.13
w3lib==2.1.1
websockets==10.4
zipp==3.11.0
